Separate selectors file
